requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
itsdangerous==2.2.0
python-dotenv==1.0.1
jinja2==3.1.4